+++ tests/collision_test.py
@@ -1,5 +1,4 @@
 from typing import List
-from tdw.tdw_utils import TDWUtils
 from sticky_mitten_avatar import StickyMittenAvatarController
 from sticky_mitten_avatar.task_status import TaskStatus
 
@@ -14,9 +13,7 @@ class CollisionTest(StickyMittenAvatarController):
         self.o_id = 0
 
     def _get_scene_init_commands(self, scene: str = None, layout: int = None, room: int = -1) -> List[dict]:
-        commands = [{"$type": "load_scene",
-                     "scene_name": "ProcGenScene"},
-                    TDWUtils.create_empty_room(4, 4)]
+        commands = super()._get_scene_init_commands()
         self.o_id, o_commands = self._add_object("trunck",
                                                  position={"x": 0, "y": 0, "z": 1})
         commands.extend(o_commands)
@@ -29,7 +26,13 @@ if __name__ == "__main__":
     # Crash into the object and stop.
     result = c.go_to(target=c.o_id)
     assert result == TaskStatus.collided_with_something_heavy, result
-    # Crash into a wall and stop.
+    # Crash into the object and stop.
     result = c.go_to(target={"x": 1.85, "y": 0, "z": -0.36})
+    assert result == TaskStatus.collided_with_something_heavy, result
+    # Ignore collisions.
+    result = c.go_to(target={"x": 0.85, "y": 0, "z": -0.36}, stop_on_collision=False)
+    assert result == TaskStatus.success, result
+    # Crash into a wall and stop.
+    result = c.go_to(target={"x": 13.85, "y": 0, "z": -0.36})
     assert result == TaskStatus.collided_with_environment, result
     c.end()